Remove async attribute
The `async` and `defer` attributes should not be used together when loading a script. Since `async` [has a higher priority](https://stackoverflow.com/a/55389033/9046809), `defer` will be ignored.

I advise to remove `async` to make proper use of `defer`, which will not block the render process at all, as shown in [this](https://stackoverflow.com/a/39711009/9046809) thread.